picfectapi/requirements.txt
astroid==1.4.7
backports.functools-lru-cache==1.2.1
configparser==3.5.0
coverage==4.2
coveralls==1.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-braces==1.9.0
django-cors-headers==1.1.0
django-envie==0.0.9
django-filter==0.13.0
django-nose==1.4.4
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.4.0
docopt==0.6.2
drfdocs==0.0.11
flake8==2.6.2
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.6
mccabe==0.5.0
nose==1.3.7
oauthlib==1.0.3
Pillow==3.3.0
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
PyJWT==1.4.1
pylint==1.6.4
python-openid==2.2.5
python-social-auth==0.2.19
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
static3==0.7.0
wrapt==1.10.8
